refactor: Rest Client 모듈 생성 및 관련 로직 분리
[#64]
