+++ query/src/main/kotlin/com/hs/application/usecase/ProductAggregateCommand.kt
@@ -1,8 +1,8 @@
 package com.hs.application.usecase
 
+import com.hs.RestGetRequestor
 import com.hs.dto.FindProductDto
 import com.hs.entity.ProductAggregate
-import com.hs.infrastructure.rest.RestGetRequestor
 import com.hs.vo.ProductInfo
 import com.hs.repository.QueryAppProductAggregateRepository
 import kotlinx.coroutines.Deferred
@@ -21,12 +21,7 @@ class ProductAggregateCommand(
 
     @Caching(
         evict = [
-            CacheEvict(key = "#productId", cacheResolver = "productAggregateCacheEvictResolver"),
-//            CacheEvict(
-//                value = ["productAggregatePage"],
-//                key = "#productId",
-//                cacheResolver = "productAggregatePageCacheEvictResolver"
-//            )
+            CacheEvict(key = "#productId", cacheResolver = "productAggregateCacheEvictResolver")
         ]
     )
     fun createOrUpdate(productId: Long) = runBlocking {
